+++ fiddles/bash/fiddle-0092-InitJavaProject/run.sh
@@ -1,7 +1,5 @@
 #!/usr/bin/env bash
 
-source bin/_utils.sh;
-
 clear;
 echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
 echo "Bash version ${BASH_VERSION}..."
@@ -9,7 +7,7 @@ echo "Bash version ${BASH_VERSION}..."
 
 function catch() {
     case $1 in
-        0)  endLog "environment configured";
+        0)  echo "";
             ;;
         *)  echo "fubar! Something went wrong."
             ;;
@@ -18,9 +16,9 @@ function catch() {
 }
 # try
 (
-    _path="fiddle";
-    if [ "$#" -eq 1 ]; then _path=$1; fi
-    # Do stuff
+    _path="fiddle-0001-Console";
+    cd scripts;
 
+    ./fiddle-java.sh ${_path};
 )
 catch $?;