+++ fiddles/nativeScript/fiddle-0003-RssReader/rssreader/app/shared/config.js
@@ -1,0 +1,4 @@
+module.exports = {
+	rssURL:"http://feeds.feedburner.com/mollythebedlingtonterrier",
+	title:"Fiddle #3 - RSS Reader"
+}